+++ ChangeLog.md
@@ -11,6 +11,7 @@
 * The `-o` parameter of the command `project GetLicenseInfo` is now optional.
   But you still need this output when you want to create a Readme.
 * `project createbom` uses source url and purl from SW360 if available
+* `project createbom` adds first SW360 source attachment of each releases as external reference to SBOM.
 * `project createbom` adds SW360 project name, version and description to SBOM.
 
 ## 2.0.0 (2023-06-02)